requirements.txt
argparse==1.2.1
paho-mqtt==1.1
requests==2.12.3
wsgiref==0.1.2
